+++ src/main/java/com/kh/eco/comment/model/service/CommentService.java
@@ -7,6 +7,6 @@ import com.kh.eco.comment.model.dto.CommentReportDTO;
 public interface CommentService {
     int insertComment(CommentDTO comment);
     int updateComment(CommentDTO comment, CustomUserDetails user);
-    int deleteComment(int commentNo, CustomUserDetails user);
+    int deleteComment(Long commentNo, CustomUserDetails user); // Long 확인
     int reportComment(CommentReportDTO report, CustomUserDetails user);
 }
